Move test code to datastructures_test.py
